meta/mtemplate: add `ModuleTemplate` as a named `ModuleDeclSet`
As an eventual replacement of the `Module` sort, `ModuleTemplate` is just a named `ModuleDeclSet`.
When printed, `Module` and `ModuleTemplate` look the same, but `ModuleTemplate` uses the more flexible `ModuleDeclSet` data-structures.

This is in opposition to the `Module` data-structures, which act as a large tuple instead.
To recover the tuple view of it, you can use explicit matching using the subsorts of `ModuleDeclSet`.