tests/maude-test-wrapper: Add infrastructure for conditionally running tests
Some tests need to be run conditionally (e.g. tests/Misc/smtTest.maude needs to
be run only if maude has been compiled with an SMT solver.) For these cases, if
the "<test_name>.check" exists, we source it. If it returns non-zero we pass
that on and let automakes test infrastructure interpret the exit code.

We source this file instead of running it so that we can define helper functions
for it to use. Currently this is a single function `check_config_h` for checking
if macros are defined in `config.h`, and a variable `exit_code_skip` that
automake's test infrastructure interprets as a singal to skip the test.

This infrastructure will be needed for the Nelson-Oppen tests that require a
particular SMT engine.