meta/eqform: fix subsorting bug
Fix bug in NNF/NEF/CNF/DNF by converting unconditional equations to
conditional equations to forbid the appearance of unwanted identities
underneath /\ and \/ when recursing. A better solution is to modify the
subsort hierarchy so that the sort system explicitly forbids the terms
in question.

Note: some bug still present in NEF.