varsat/: update to newest var-sat tool
- Rework dependencies underneath varsat; thus new versions of
dependencies should generally be prefered compared to older versions;
may not hold true for FOFORM since you guys fixed bugs there.

- (Almost totally) remove FULL-MAUDE dependency except for
SUBSTITUTIONSET and maybe a few other things --- in particular, see
UNIT-FM in meta-aux for a UNIT clone that does most of what the
FULL-MAUDE version does except that it does not support the 'omod'
syntax.