rltool: remove comment to investigate
NOTE: the constraint can become false when the pattern difference
overapproximation succeeds and the precondition and postcondition
have the same term and constraint (but renamed). That is, when we
have goals of the form:

(t | φ) ⇒ (t' | φ')

such that (t' | φ') is a renaming of (t | φ). Then pattern difference
overapproximation gives us: (t | φ ∧ ¬φ) which our tool is smart enough
to reduce to (t | ⊥) which the subsumption check immediately discharges.