full-maude: add removeIds full subsumption feature flag
Add boolean flag to control removeIds subsumption behavior.
If flag is true, perform full equation/rule subsumption.
Otherwise, only check that variants of a single equation or rule are
subsumed with respect to other variants of the same equation/rule.
In practice, this process generates a sufficiently small number of
equations/rules.